disk: move PartitionTable and methods to separate file
Split up disk.go into smaller more managable pieces: start with
extracting `PartitionTable` and all its methods to a new file.
